eval/COMET/requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
filelock==3.0.12
fsspec==2021.7.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.4
joblib==1.0.1
jsonargparse==3.13.1
Markdown==3.3.4
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.1.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.3.5
pytz==2021.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tokenizers==0.10.3
torchmetrics==0.5.0
tqdm==4.62.2
transformers==4.9.2
typing-extensions==3.10.0.0
unbabel-comet==1.0.0rc4
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
